pyproject.toml
[tool.poetry]
name = "dolphin-lit-rm"
version = "0.1.0"
description = "Data pipeline for Literary Reward Model training."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0"
typer = {extras = ["all"], version = "^0.9.0"}
requests = "^2.31.0"
pyarrow = "^14.0.1"
datasets = "^2.15.0" # Hugging Face datasets
loguru = "^0.7.2"
fasttext = "^0.9.2"
spacy = "^3.7.2"
# For spacy model: python -m spacy download en_core_web_sm
lmdb = "^1.4.1"
jinja2 = "^3.1.2"
tqdm = "^4.66.1"
tiktoken = "^0.5.1" # For token counting

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dolphin-lit-rm = "dolphin_lit_rm.cli:app"